Moved geometry helper functions to polygonSlice.ts
